refactor: RepositoryMapper 에서 EntityManagerManager로 역할 위임
